+++ src/test/java/crypto/CryptoResponseTest.java
@@ -10,14 +10,12 @@ import org.junit.Test;
 
 import util.TestUtils;
 import static util.TestUtils.json;
-import static util.TestUtils.stream;
+import static util.TestUtils.error;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
+import static org.junit.Assert.assertNotEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertTrue;
-import static util.TestUtils.error;
-import static util.TestUtils.errorMessage;
-import static util.TestUtils.exchangeRateUrl;
 
 public class CryptoResponseTest {
 
@@ -32,6 +30,14 @@ public class CryptoResponseTest {
         assertNotNull(response.getMetaData());
         assertEquals(response.getCryptoUnits().size(), 2);
         assertTrue(response.toString().matches("(.*), errorMessage='null'(.*)"));
+        assertNotEquals(response.getMetaData().getDigitalCurrencyCode(), "");
+        assertNotEquals(response.getMetaData().getDigitalCurrencyName(), "");
+        assertNotEquals(response.getMetaData().getInformation(), "");
+        assertNotEquals(response.getMetaData().getLastRefreshed(), "");
+        assertNotEquals(response.getMetaData().getMarketCode(), "");
+        assertNotEquals(response.getMetaData().getMarketName(), "");
+        assertNotEquals(response.getMetaData().getTimeZone(), "");
+
     }
     
     @Test