+++ src/test/java/crypto/CryptoTest.java
@@ -1,22 +1,16 @@
 package crypto;
 
-import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
-import static org.junit.Assert.assertTrue;
+import static util.TestUtils.cryptoRatingUrl;
+import static util.TestUtils.cryptoUrl;
+import static util.TestUtils.errorMessage;
+import static util.TestUtils.stream;
 
 import java.io.IOException;
 import java.util.concurrent.CountDownLatch;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicReference;
-
-import util.TestUtils;
-import static util.TestUtils.json;
-import static util.TestUtils.stream;
-import static util.TestUtils.error;
-import static util.TestUtils.errorMessage;
-import static util.TestUtils.cryptoRatingUrl;
-import static util.TestUtils.cryptoUrl;
 
 import com.crazzyghost.alphavantage.AlphaVantage;
 import com.crazzyghost.alphavantage.AlphaVantageException;
@@ -33,6 +27,7 @@ import okhttp3.logging.HttpLoggingInterceptor;
 import okhttp3.logging.HttpLoggingInterceptor.Level;
 import okhttp3.mock.Behavior;
 import okhttp3.mock.MockInterceptor;
+import util.TestUtils;
 
 public class CryptoTest {